Fix migration table and room modal

+++ database/migrations/2024_11_14_115519_remove_course_id_from_schedules_table.php
@@ -4,33 +4,22 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class RemoveCourseIdFromSchedulesTable extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up()
     {
         Schema::table('schedules', function (Blueprint $table) {
-
-            $table->dropForeign(['course_id']);
-
-
-            $table->dropColumn('course_id');
+            if (Schema::hasColumn('schedules', 'course_id')) {
+                $table->dropForeign(['course_id']);
+                $table->dropColumn('course_id');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down()
     {
         Schema::table('schedules', function (Blueprint $table) {
-
-            $table->unsignedBigInteger('course_id')->nullable();
-
-
-            $table->foreign('course_id')->references('id')->on('courses')->onDelete('cascade');
+            $table->foreignId('course_id')->nullable()->constrained()->onDelete('cascade');
         });
     }
-};
+}
